+++ src/store/modules/navigator/mutations_types.js
@@ -1,0 +1,6 @@
+export const PUSH = 'PUSH'
+export const POP = 'POP'
+export const REPLACE = 'REPLACE'
+export const RESET = 'RESET'
+export const OPTIONS = 'OPTIONS'
+export const SET = 'SET'